refactor: 인수 테스트 static 메서드 분리
